requirements.txt
# Core dependencies for package metadata extraction
requests>=2.28.0
lxml>=4.9.0
cyclonedx-python-lib[validation]>=11.0.0

# Note: The GUI (gui_menu.py) requires tkinter, which is included 
# with most Python installations by default. If tkinter is not 
# available, you can still use the command-line interface.